OpenPiton Release 9 (18-11-20-r9)
These are all the changes related to the integration of the PicoRV32
core with the OpenPiton framework to create "JuxtaPiton". They enable
the instantiation of the framework using OpenPiton and PicoRV32 cores
in simulation and on FPGA. Designs can be implemented with just SPARC
cores, just PicoRV32 cores, or a mix. Changes include:
-Adding the PicoRV32 core to the hierarchy
-Adding two shims (pico_decoder and l15_picoencoder) that sit
between the PicoRV32 core and the L1.5 cache that enable PicoRV32 to
interface with the OpenPiton cache system
-Modifications to various instantiations throughout tile and L1.5 to
instantiate PicoRV32 or OpenPiton structures as appropriate
-Modifications to infrastructure files for FPGA and simulation to
configure the instantiation of PicoRV32 or OpenPiton cores

---

The changes also add support for running riscv assembly tests in simulation.

The tests used are the same as those provided with picorv32, rather
than the ones provided in riscv-tests, but I think the difference
is minor.

This passes all of the provided assembly tests. You can run them
using `sims -group=pico_tile1 -sim_type=vcs` (replacing vcs with
icv/ncv/msm if you prefer a different compiler).

If you want to add a new test, throw it in
`piton/verif/diag/assembly/riscv/pico/` and make sure you use a
capital S for the test filename's extension.

This isn't intended to be robust to arbitrary tests you write. A
later commit should at least support the C benchmarks that come
as part of riscv-tests.

Small issue: simple.S actually doesn't pass but that's just some
because there's no code produced for the pass or fail labels.

---

The final thing included is support for RISC-V AMO ops (but not
LR/SC) which are performed in the L2 cache.